refactor(solobase-native): extract register_http_listener + serve_until_shutdown
